Moving code around.
