+++ .travis.yml
@@ -1,0 +1,7 @@
+language: scala
+
+jdk: openjdk11
+
+script:
+  - sbt scripted
+
