+++ app/org/maproulette/controllers/api/TaskController.scala
@@ -10,10 +10,11 @@ import javax.inject.Inject
 import org.maproulette.Config
 import org.maproulette.controllers.CRUDController
 import org.maproulette.data._
-import org.maproulette.models.dal.{TagDAL, TagDALMixin, TaskDAL, DALManager}
+import org.maproulette.models.dal.{DALManager, TagDAL, TaskDAL}
 import org.maproulette.models._
-import org.maproulette.exception.{InvalidException, NotFoundException, LockedException, StatusMessage}
-import org.maproulette.session.{SearchLocation, SearchParameters, SearchChallengeParameters, SessionManager, User}
+import org.maproulette.exception.{InvalidException, LockedException, NotFoundException, StatusMessage}
+import org.maproulette.models.dal.mixin.TagDALMixin
+import org.maproulette.session.{SearchChallengeParameters, SearchLocation, SearchParameters, SessionManager, User}
 import org.maproulette.utils.Utils
 import org.maproulette.services.osm._
 import org.maproulette.provider.websockets.{WebSocketMessages, WebSocketProvider}
@@ -24,7 +25,7 @@ import play.api.libs.ws.WSClient
 import play.api.mvc._
 
 import scala.concurrent.duration._
-import scala.concurrent.{Await, Promise, Future}
+import scala.concurrent.{Await, Future, Promise}
 import scala.util.{Failure, Success}
 
 /**
@@ -58,6 +59,7 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
 
   // The type of object that this controller deals with.
   override implicit val itemType = TaskType()
+  override implicit val tableName = this.dal.tableName
   // json reads for automatically reading Tags from a posted json body
   implicit val tagReads: Reads[Tag] = Tag.tagReads
   implicit val commentReads: Reads[Comment] = Comment.commentReads
@@ -320,7 +322,7 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
     }
 
     val tags = tagDAL.listByTask(taskToReturn.id)
-    return Utils.insertIntoJson(Json.toJson(taskToReturn), Tag.KEY, Json.toJson(tags.map(_.name)))
+    Utils.insertIntoJson(Json.toJson(taskToReturn), Tag.KEY, Json.toJson(tags.map(_.name)))
   }
 
   /**
@@ -340,7 +342,7 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
     this.sessionManager.userAwareRequest { implicit user =>
       SearchParameters.withSearch { p =>
         val params = p.copy(location = Some(SearchLocation(left, bottom, right, top)))
-        val (count, result) = this.dal.getTasksInBoundingBox(User.userOrMocked(user), params, limit, offset,
+        val (count, result) = this.dalManager.taskCluster.getTasksInBoundingBox(User.userOrMocked(user), params, limit, offset,
                                 excludeLocked, sort, order)
 
         val resultJson = _insertExtraJSON(result, includeGeometries)
@@ -381,55 +383,6 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
     }
   }
 
-  /**
-    * This performs setTaskStatus on a bundle of tasks.
-    *
-    * @param bundleId     The id of the task bundle
-    * @param primaryId    The id of the primary task for this bundle
-    * @param status The status id to set the task's status to
-    * @param comment An optional comment to add to the task
-    * @param tags Optional tags to add to the task
-    * @return 400 BadRequest if status id is invalid or task with supplied id not found.
-    *         If successful then 200 NoContent
-    */
-  def setBundleTaskStatus(bundleId: Long, primaryId: Long, status: Int, comment: String = "", tags: String = ""): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      val requestReview = request.getQueryString("requestReview") match {
-        case Some(v) => Some(v.toBoolean)
-        case None => None
-      }
-
-      val tasks = this.dal.getTaskBundle(user, bundleId).tasks match {
-        case Some(t) => t
-        case None => throw new InvalidException("No tasks found in this bundle.")
-      }
-
-      val completionResponses = request.body.asJson
-      this.dal.setTaskStatus(tasks, status, user, requestReview, completionResponses, Some(bundleId), Some(primaryId))
-
-      for (task <- tasks) {
-        val action = this.actionManager.setAction(Some(user), new TaskItem(task.id), TaskStatusSet(status), task.name)
-        // add comment to each task if any provided
-        if (comment.nonEmpty) {
-          val actionId = action match {
-            case Some(a) => Some(a.id)
-            case None => None
-          }
-          this.dal.addComment(user, task, comment, actionId)
-        }
-
-        // Add tags to each task
-        val tagList = tags.split(",").toList
-        if (tagList.nonEmpty) {
-          this.addTagstoItem(task.id, tagList.map(new Tag(-1, _, tagType = this.dal.tableName)), user)
-        }
-      }
-
-      // Refetch to get updated data
-      Ok(Json.toJson(this.dal.getTaskBundle(user, bundleId)))
-    }
-  }
-
   def customTaskStatus(taskId:Long, actionType: ActionType, user:User, comment:String= "",
                        tags: String= "",requestReview:Option[Boolean] = None, completionResponses:Option[JsValue] = None) = {
     val status = actionType match {
@@ -455,7 +408,7 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
         case Some(a) => Some(a.id)
         case None => None
       }
-      this.dal.addComment(user, task, comment, actionId)
+      this.dalManager.comment.add(user, task, comment, actionId)
     }
 
     val tagList = tags.split(",").toList
@@ -489,7 +442,7 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
         case None => None
       }
 
-      this.dal.setTaskReviewStatus(task, reviewStatus, user, actionId, comment)
+      this.dalManager.taskReview.setTaskReviewStatus(task, reviewStatus, user, actionId, comment)
 
       val tagList = tags.split(",").toList
       if (tagList.nonEmpty) {
@@ -497,45 +450,6 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
       }
 
       NoContent
-    }
-  }
-
-  /**
-    * This function sets the task review status.
-    * Must be authenticated to perform operation and marked as a reviewer.
-    *
-    * @param id The id of the task
-    * @param reviewStatus The review status id to set the task's review status to
-    * @param comment An optional comment to add to the task
-    * @param tags Optional tags to add to the task
-    * @return 400 BadRequest if task with supplied id not found.
-    *         If successful then 200 NoContent
-    */
-  def setBundleTaskReviewStatus(id: Long, reviewStatus: Int, comment:String= "", tags: String= "") : Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      val tasks = this.dal.getTaskBundle(user, id).tasks match {
-        case Some(t) => t
-        case None => throw new InvalidException("No tasks found in this bundle.")
-      }
-
-      for (task <- tasks) {
-        val action = this.actionManager.setAction(Some(user), new TaskItem(task.id),
-                       TaskReviewStatusSet(reviewStatus), task.name)
-        val actionId = action match {
-          case Some(a) => Some(a.id)
-          case None => None
-        }
-
-        this.dal.setTaskReviewStatus(task, reviewStatus, user, actionId, comment)
-
-        val tagList = tags.split(",").toList
-        if (tagList.nonEmpty) {
-          this.addTagstoItem(id, tagList.map(new Tag(-1, _, tagType = this.dal.tableName)), user)
-        }
-      }
-
-      // Refetch to get updated data
-      Ok(Json.toJson(this.dal.getTaskBundle(user, id)))
     }
   }
 
@@ -561,101 +475,6 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
   }
 
   /**
-    * Retrieves a specific comment for the user
-    *
-    * @param commentId The id of the comment to retrieve
-    * @return The comment
-    */
-  def retrieveComment(commentId: Long): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.userAwareRequest { implicit user =>
-      this.dal.retrieveComment(commentId) match {
-        case Some(comment) => Ok(Json.toJson(comment))
-        case None => NotFound
-      }
-    }
-  }
-
-  /**
-    * Retrieves all the comments for a Task
-    *
-    * @param taskId The task to retrieve the comments for
-    * @return A list of comments
-    */
-  def retrieveComments(taskId: Long): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.userAwareRequest { implicit user =>
-      Ok(Json.toJson(this.dal.retrieveComments(List.empty, List.empty, List(taskId))))
-    }
-  }
-
-  /**
-    * Adds a comment for a specific task
-    *
-    * @param taskId   The id for a task
-    * @param comment  The comment the user is leaving
-    * @param actionId The action if any associated with the comment
-    * @return Ok if successful.
-    */
-  def addComment(taskId: Long, comment: String, actionId: Option[Long]): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      val task = this.dal.retrieveById(taskId) match {
-        case Some(t) => t
-        case None => throw new NotFoundException(s"Task with $taskId not found, can not add comment.")
-      }
-      Created(Json.toJson(this.dal.addComment(user, task, URLDecoder.decode(comment, "UTF-8"), actionId)))
-    }
-  }
-
-  /**
-    * Adds a comment for tasks in a bundle
-    *
-    * @param bundleId   The id for the bundle
-    * @param comment  The comment the user is leaving
-    * @param actionId The action if any associated with the comment
-    * @return Ok if successful.
-    */
-  def addCommentToBundleTasks(bundleId: Long, comment: String, actionId: Option[Long]): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      val tasks = this.dal.getTaskBundle(user, bundleId).tasks match {
-        case Some(t) => t
-        case None => throw new InvalidException("No tasks found in this bundle.")
-      }
-
-      for (task <- tasks) {
-        this.dal.addComment(user, task, URLDecoder.decode(comment, "UTF-8"), actionId)
-      }
-
-      Ok(Json.toJson(this.dal.getTaskBundle(user, bundleId)))
-    }
-  }
-
-  /**
-    * Updates the original comment
-    *
-    * @param commentId The ID of the comment to update
-    * @param comment   The comment to update
-    * @return
-    */
-  def updateComment(commentId: Long, comment: String): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      Ok(Json.toJson(this.dal.updateComment(user, commentId, URLDecoder.decode(comment, "UTF-8"))))
-    }
-  }
-
-  /**
-    * Deletes a comment from a task
-    *
-    * @param taskId    The id of the task that the comment is associated with
-    * @param commentId The id of the comment that is being deleted
-    * @return Ok if successful,
-    */
-  def deleteComment(taskId: Long, commentId: Long): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      this.dal.deleteComment(user, taskId, commentId)
-      Ok
-    }
-  }
-
-  /**
     * Gets clusters of tasks for the challenge. Uses kmeans method in postgis.
     *
     * @param numberOfPoints Number of clustered points you wish to have returned
@@ -664,7 +483,7 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
   def getTaskClusters(numberOfPoints: Int): Action[AnyContent] = Action.async { implicit request =>
     this.sessionManager.userAwareRequest { implicit user =>
       SearchParameters.withSearch { implicit params =>
-        Ok(Json.toJson(this.dal.getTaskClusters(params, numberOfPoints)))
+        Ok(Json.toJson(this.dalManager.taskCluster.getTaskClusters(params, numberOfPoints)))
       }
     }
   }
@@ -681,7 +500,7 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
   def getTasksInCluster(clusterId: Int, numberOfPoints: Int): Action[AnyContent] = Action.async { implicit request =>
     this.sessionManager.userAwareRequest { implicit user =>
       SearchParameters.withSearch { implicit params =>
-        Ok(Json.toJson(this.dal.getTasksInCluster(clusterId, params, numberOfPoints)))
+        Ok(Json.toJson(this.dalManager.taskCluster.getTasksInCluster(clusterId, params, numberOfPoints)))
       }
     }
   }
@@ -720,63 +539,6 @@ class TaskController @Inject()(override val sessionManager: SessionManager,
           p.future
         }
       )
-    }
-  }
-
-  /**
-    * Creates a new task bundle with the task ids in the json body, assigning
-    * ownership of the bundle to the logged-in user
-    *
-    * @return A TaskBundle representing the new bundle
-    */
-  def createTaskBundle(): Action[JsValue] = Action.async(bodyParsers.json) { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      val name = (request.body \ "name").asOpt[String].getOrElse("")
-      val taskIds = (request.body \ "taskIds").asOpt[List[Long]] match {
-        case Some(tasks) => tasks
-        case None => throw new InvalidException("No task ids provided for task bundle")
-      }
-      val bundle = dal.createTaskBundle(user, name, taskIds)
-      Created(Json.toJson(bundle))
-    }
-  }
-
-  /**
-    * Gets the tasks in the given Bundle
-    *
-    * @param id
-    * @return Task Bundle
-    */
-  def getTaskBundle(id: Long): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      Ok(Json.toJson(this.dal.getTaskBundle(user, id)))
-    }
-  }
-
-  /**
-    * Remove tasks from a bundle.
-    *
-    * @param id
-    * @param taskIds List of task ids to remove
-    * @return Task Bundle
-    */
-  def unbundleTasks(id: Long, taskIds: List[Long]): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      this.dal.unbundleTasks(user, id, taskIds)
-      Ok(Json.toJson(this.dal.getTaskBundle(user, id)))
-    }
-  }
-
-  /**
-    * Delete bundle.
-    *
-    * @param id
-    * @param primaryId optional task id to no unlcok after deleting this bundle
-    */
-  def deleteTaskBundle(id: Long, primaryId: Option[Long] = None): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.authenticatedRequest { implicit user =>
-      this.dal.deleteTaskBundle(user, id, primaryId)
-      Ok
     }
   }
