+++ app/org/maproulette/models/dal/TaskReviewDAL.scala
@@ -3,56 +3,91 @@
 package org.maproulette.models.dal
 
 import java.sql.Connection
+import java.time.Instant
 
+import anorm.JodaParameterMetaData._
 import anorm.SqlParser._
 import anorm._
 import javax.inject.{Inject, Provider, Singleton}
-import org.joda.time.{DateTime, DateTimeZone}
+import org.joda.time.DateTime
+import org.maproulette.Config
 import org.maproulette.exception.InvalidException
 import org.maproulette.models._
-import org.maproulette.data._
-import org.maproulette.Config
 import org.maproulette.permissions.Permission
-import org.maproulette.session.{User, SearchParameters}
-import org.maproulette.session.dal.UserDAL
-import org.maproulette.provider.websockets.WebSocketProvider
-import org.maproulette.provider.websockets.WebSocketMessages
+import org.maproulette.provider.websockets.{WebSocketMessages, WebSocketProvider}
+import org.maproulette.session.{SearchParameters, User}
 import play.api.db.Database
 import play.api.libs.ws.WSClient
-import play.api.libs.json._
 
+import scala.collection.mutable
 import scala.collection.mutable.ListBuffer
-import scala.collection.mutable.Map
 
 /**
   * @author krotstan
   */
 @Singleton
 class TaskReviewDAL @Inject()(override val db: Database,
-                                override val tagDAL: TagDAL, config: Config,
-                                override val permission: Permission,
-                                taskDAL: TaskDAL,
-                                userDAL: Provider[UserDAL],
-                                projectDAL: Provider[ProjectDAL],
-                                challengeDAL: Provider[ChallengeDAL],
-                                notificationDAL: Provider[NotificationDAL],
-                                actions: ActionManager,
-                                statusActions: StatusActionManager,
-                                webSocketProvider: WebSocketProvider,
-                                ws: WSClient)
-  extends TaskDAL(db, tagDAL, config, permission, userDAL, projectDAL, challengeDAL, notificationDAL,
-                  actions, statusActions, webSocketProvider, ws) {
+                              override val tagDAL: TagDAL,
+                              override val permission: Permission,
+                              config: Config,
+                              dalManager: Provider[DALManager],
+                              webSocketProvider: WebSocketProvider,
+                              ws: WSClient)
+  extends TaskDAL(db, tagDAL, permission, config, dalManager, webSocketProvider, ws) {
+
+  implicit val taskWithReviewParser: RowParser[TaskWithReview] = {
+    get[Long]("tasks.id") ~
+      get[String]("tasks.name") ~
+      get[DateTime]("tasks.created") ~
+      get[DateTime]("tasks.modified") ~
+      get[Long]("parent_id") ~
+      get[Option[String]]("tasks.instruction") ~
+      get[Option[String]]("geo_location") ~
+      get[Option[Int]]("tasks.status") ~
+      get[Option[String]]("geo_json") ~
+      get[Option[String]]("suggested_fix") ~
+      get[Option[DateTime]]("tasks.mapped_on") ~
+      get[Option[Int]]("task_review.review_status") ~
+      get[Option[Long]]("task_review.review_requested_by") ~
+      get[Option[Long]]("task_review.reviewed_by") ~
+      get[Option[DateTime]]("task_review.reviewed_at") ~
+      get[Option[DateTime]]("task_review.review_started_at") ~
+      get[Option[Long]]("task_review.review_claimed_by") ~
+      get[Int]("tasks.priority") ~
+      get[Option[Long]]("tasks.changeset_id") ~
+      get[Option[Long]]("tasks.bundle_id") ~
+      get[Option[Boolean]]("tasks.is_bundle_primary") ~
+      get[Option[String]]("challenge_name") ~
+      get[Option[String]]("review_requested_by_username") ~
+      get[Option[String]]("reviewed_by_username") ~
+      get[Option[String]]("responses") map {
+      case id ~ name ~ created ~ modified ~ parent_id ~ instruction ~ location ~ status ~ geojson ~
+        suggestedFix ~ mappedOn ~ reviewStatus ~ reviewRequestedBy ~ reviewedBy ~ reviewedAt ~
+        reviewStartedAt ~ reviewClaimedBy ~ priority ~ changesetId ~ bundleId ~ isBundlePrimary ~
+        challengeName ~ reviewRequestedByUsername ~ reviewedByUsername ~ responses =>
+
+        val values = this.updateAndRetrieve(id, geojson, location, suggestedFix)
+        TaskWithReview(
+          Task(id, name, created, modified, parent_id, instruction, values._2,
+            values._1, values._3, status, mappedOn,
+            TaskReviewFields(reviewStatus, reviewRequestedBy, reviewedBy, reviewedAt, reviewStartedAt,
+              reviewClaimedBy), priority, changesetId, responses, bundleId, isBundlePrimary),
+          TaskReview(-1, id, reviewStatus, challengeName, reviewRequestedBy, reviewRequestedByUsername,
+            reviewedBy, reviewedByUsername, reviewedAt, reviewStartedAt, reviewClaimedBy, None, None)
+        )
+    }
+  }
 
   /**
     * Gets and claims a task for review.
     *
-    * @param user The user executing the request
-    * @param id id of task that you wish to start/claim
+    * @param user        The user executing the request
+    * @param primaryTask id of task that you wish to start/claim
     * @return task
     */
-  def startTaskReview(user:User, primaryTask:Task) (implicit c:Connection=null) : Option[Task] = {
+  def startTaskReview(user: User, primaryTask: Task)(implicit c: Connection = null): Option[Task] = {
     if (primaryTask.review.reviewClaimedBy.getOrElse(null) != null &&
-        primaryTask.review.reviewClaimedBy.getOrElse(null) != user.id.toLong) {
+      primaryTask.review.reviewClaimedBy.getOrElse(null) != user.id.toLong) {
       throw new InvalidException("This task is already being reviewed by someone else.")
     }
 
@@ -60,7 +95,7 @@ class TaskReviewDAL @Inject()(override val db: Database,
     if (primaryTask.isBundlePrimary.getOrElse(false)) {
       primaryTask.bundleId match {
         case Some(bId) =>
-          this.getTaskBundle(user, bId).tasks match {
+          this.manager.taskBundle.getTaskBundle(user, bId).tasks match {
             case Some(tList) =>
               taskList = tList
             case None => // do nothing -- just use our current task
@@ -75,8 +110,7 @@ class TaskReviewDAL @Inject()(override val db: Database,
         SQL"""UPDATE task_review SET review_claimed_by = NULL, review_claimed_at = NULL
                 WHERE review_claimed_by = #${user.id}""".executeUpdate()
 
-        val updatedRows =
-          SQL"""UPDATE task_review SET review_claimed_by = #${user.id}, review_claimed_at = NOW()
+        SQL"""UPDATE task_review SET review_claimed_by = #${user.id}, review_claimed_at = NOW()
                   WHERE task_id = #${task.id} AND review_claimed_at IS NULL""".executeUpdate()
 
         try {
@@ -86,10 +120,10 @@ class TaskReviewDAL @Inject()(override val db: Database,
         }
 
         implicit val id = task.id
-        this.taskDAL.cacheManager.withUpdatingCache(Long => this.retrieveById) { implicit cachedItem =>
-            val result = Some(task.copy(review = task.review.copy(reviewClaimedBy = Option(user.id.toInt))))
-              result
-            }(task.id, true, true)
+        this.manager.task.cacheManager.withUpdatingCache(Long => this.retrieveById) { implicit cachedItem =>
+          val result = Some(task.copy(review = task.review.copy(reviewClaimedBy = Option(user.id.toInt))))
+          result
+        }(task.id, true, true)
       }
     }
 
@@ -108,17 +142,18 @@ class TaskReviewDAL @Inject()(override val db: Database,
     * Releases a claim on a task for review.
     *
     * @param user The user executing the request
-    * @param id id of task that you wish to release
+    * @param task id of task that you wish to release
     * @return task
     */
-  def cancelTaskReview(user:User, task:Task) (implicit c:Connection=null) : Option[Task] = {
+  def cancelTaskReview(user: User, task: Task)(implicit c: Connection = null): Option[Task] = {
     if (task.review.reviewClaimedBy.getOrElse(null) != user.id.toLong) {
       throw new InvalidException("This task is not currently being reviewed by you.")
     }
 
     this.withMRTransaction { implicit c =>
-      val updatedRows = SQL"""UPDATE task_review t SET review_claimed_by = NULL, review_claimed_at = NULL
-              WHERE t.task_id = #${task.id}""".executeUpdate()
+      val updatedRows =
+        SQL"""UPDATE task_review SET review_claimed_by = NULL, review_claimed_at = NULL
+              WHERE task_review.task_id = #${task.id}""".executeUpdate()
 
       // if returning 0, then this is because the item is locked by a different user
       if (updatedRows == 0) {
@@ -137,15 +172,34 @@ class TaskReviewDAL @Inject()(override val db: Database,
     }
 
     val updatedTask = task.copy(review = task.review.copy(reviewClaimedBy = None))
-    this.taskDAL.cacheManager.withOptionCaching { () => Some(updatedTask) }
+    this.manager.task.cacheManager.withOptionCaching { () => Some(updatedTask) }
     Option(updatedTask)
+  }
+
+  def getTaskWithReview(taskId: Long): TaskWithReview = {
+    this.withMRConnection { implicit c =>
+      val query =
+        s"""
+        SELECT $retrieveColumnsWithReview,
+               challenges.name as challenge_name,
+               mappers.name as review_requested_by_username,
+               reviewers.name as reviewed_by_username
+        FROM ${this.tableName}
+        LEFT OUTER JOIN task_review ON task_review.task_id = tasks.id
+        LEFT OUTER JOIN users mappers ON task_review.review_requested_by = mappers.id
+        LEFT OUTER JOIN users reviewers ON task_review.reviewed_by = reviewers.id
+        INNER JOIN challenges ON challenges.id = tasks.parent_id
+        WHERE tasks.id = {taskId}
+      """
+      SQL(query).on('taskId -> taskId).as(this.taskWithReviewParser.single)
+    }
   }
 
   /**
     * Gets and claims the next task for review.
     *
     * @param user The user executing the request
-    * @param SearchParameters
+    * @param searchParameters
     * @param onlySaved
     * @param sort
     * @param order
@@ -153,24 +207,24 @@ class TaskReviewDAL @Inject()(override val db: Database,
     * @param excludeOtherReviewers
     * @return task
     */
-  def nextTaskReview(user:User, searchParameters: SearchParameters, onlySaved: Boolean=false,
-                    sort:String, order:String, lastTaskId:Option[Long]=None, excludeOtherReviewers: Boolean=false) (implicit c:Connection=null) : Option[Task] = {
+  def nextTaskReview(user: User, searchParameters: SearchParameters, onlySaved: Boolean = false,
+                     sort: String, order: String, lastTaskId: Option[Long] = None, excludeOtherReviewers: Boolean = false)(implicit c: Connection = null): Option[Task] = {
     var position = 0
 
     lastTaskId match {
       case Some(taskId) => {
         val (countAll, queryAll, parametersAll) =
           _getReviewRequestedQueries(user, searchParameters, null, null, onlySaved,
-                                     -1, 0, sort, order, false, excludeOtherReviewers)
+            -1, 0, sort, order, false, excludeOtherReviewers)
 
         // This only happens if a non-reviewer is trying to do get a review task
         if (queryAll == null) {
           return None
         }
 
-        var rowMap = Map[Long, Int]()
+        val rowMap = mutable.Map[Long, Int]()
         val rowNumParser: RowParser[Long] = {
-            get[Int]("row_num") ~
+          get[Int]("row_num") ~
             get[Long]("tasks.id") map {
             case row ~ id => {
               rowMap.put(id, row)
@@ -185,7 +239,7 @@ class TaskReviewDAL @Inject()(override val db: Database,
           val rowPosition = rowMap.get(taskId)
           rowPosition match {
             case Some(row) =>
-              position = row  // fetch next task (offset starts at 0 but rows at 1)
+              position = row // fetch next task (offset starts at 0 but rows at 1)
             case _ => // not found so do nothing
           }
         }
@@ -195,7 +249,7 @@ class TaskReviewDAL @Inject()(override val db: Database,
 
     val (countQuery, query, parameters) =
       _getReviewRequestedQueries(user, searchParameters, null, null, onlySaved,
-                                 1, position, sort, order, false, excludeOtherReviewers)
+        1, position, sort, order, false, excludeOtherReviewers)
 
     // This only happens if a non-reviewer is trying to do get a review task
     if (query == null) {
@@ -210,24 +264,24 @@ class TaskReviewDAL @Inject()(override val db: Database,
   /**
     * Gets a list of tasks that have requested review (and are in this user's project group)
     *
-    * @param user The user executing the request
-    * @param startDate Limit tasks to reviewed after date (YYYY-MM-DD)
-    * @param endDate Limit tasks to reviewed before date (YYYY-MM-DD)
-    * @param limit The number of tasks to return
-    * @param offset Offset to start paging
-    * @param sort Sort column
-    * @param order DESC or ASC
+    * @param user            The user executing the request
+    * @param startDate       Limit tasks to reviewed after date (YYYY-MM-DD)
+    * @param endDate         Limit tasks to reviewed before date (YYYY-MM-DD)
+    * @param limit           The number of tasks to return
+    * @param offset          Offset to start paging
+    * @param sort            Sort column
+    * @param order           DESC or ASC
     * @param includeDisputed Whether disputed tasks whould be returned in results (default is true)
     * @return A list of tasks
     */
-  def getReviewRequestedTasks(user:User, searchParameters: SearchParameters,
-                              startDate:String, endDate:String, onlySaved: Boolean=false,
-                              limit:Int = -1, offset:Int=0, sort:String, order:String,
+  def getReviewRequestedTasks(user: User, searchParameters: SearchParameters,
+                              startDate: String, endDate: String, onlySaved: Boolean = false,
+                              limit: Int = -1, offset: Int = 0, sort: String, order: String,
                               includeDisputed: Boolean = true, excludeOtherReviewers: Boolean = false)
-                    (implicit c:Connection=null) : (Int, List[Task]) = {
+                             (implicit c: Connection = null): (Int, List[Task]) = {
     val (countQuery, query, parameters) =
       _getReviewRequestedQueries(user, searchParameters, startDate, endDate, onlySaved, limit,
-                                 offset, sort, order, includeDisputed, excludeOtherReviewers)
+        offset, sort, order, includeDisputed, excludeOtherReviewers)
 
     // This only happens if a non-reviewer is trying to get review tasks
     if (query == null) {
@@ -235,27 +289,25 @@ class TaskReviewDAL @Inject()(override val db: Database,
     }
 
     var count = 0
-    val tasks = this.taskDAL.cacheManager.withIDListCaching { implicit cachedItems =>
+    val tasks = this.manager.task.cacheManager.withIDListCaching { implicit cachedItems =>
       this.withMRTransaction { implicit c =>
         count = sqlWithParameters(countQuery, parameters).as(SqlParser.int("count").single)
         sqlWithParameters(query, parameters).as(this.parser.*)
       }
     }
-
-    return (count, tasks)
+    (count, tasks)
   }
 
-  def _getReviewRequestedQueries(user:User, searchParameters: SearchParameters,
-                              startDate:String, endDate:String, onlySaved: Boolean=false,
-                              limit:Int = -1, offset:Int=0, sort:String, order:String,
-                              includeDisputed: Boolean = true, excludeOtherReviewers: Boolean = false)
-                    (implicit c:Connection=null) : (String, String, ListBuffer[NamedParameter]) = {
+  def _getReviewRequestedQueries(user: User, searchParameters: SearchParameters,
+                                 startDate: String, endDate: String, onlySaved: Boolean = false,
+                                 limit: Int = -1, offset: Int = 0, sort: String, order: String,
+                                 includeDisputed: Boolean = true, excludeOtherReviewers: Boolean = false)
+                                (implicit c: Connection = null): (String, String, ListBuffer[NamedParameter]) = {
     var orderByClause = ""
-    val whereClause = includeDisputed match {
-      case true =>
-        new StringBuilder(s"(task_review.review_status=${Task.REVIEW_STATUS_REQUESTED} OR task_review.review_status=${Task.REVIEW_STATUS_DISPUTED})")
-      case false =>
-        new StringBuilder(s"task_review.review_status=${Task.REVIEW_STATUS_REQUESTED}")
+    val whereClause = if (includeDisputed) {
+      new StringBuilder(s"(task_review.review_status=${Task.REVIEW_STATUS_REQUESTED} OR task_review.review_status=${Task.REVIEW_STATUS_DISPUTED})")
+    } else {
+      new StringBuilder(s"task_review.review_status=${Task.REVIEW_STATUS_REQUESTED}")
     }
 
     val whereBundleClause = " AND (tasks.bundle_id is NULL OR tasks.is_bundle_primary = true) "
@@ -294,9 +346,8 @@ class TaskReviewDAL @Inject()(override val db: Database,
       case _ => // ignore
     }
 
-    val query = user.isSuperUser match {
-      case true =>
-        s"""
+    val query = if (user.isSuperUser) {
+      s"""
           SELECT ROW_NUMBER() OVER (${orderByClause}) as row_num,
             tasks.${this.retrieveColumnsWithReview} FROM tasks
           ${joinClause}
@@ -306,13 +357,13 @@ class TaskReviewDAL @Inject()(override val db: Database,
           ${orderByClause}
           LIMIT ${sqlLimit(limit)} OFFSET ${offset}
          """
-      case default =>
-        if (user.settings.isReviewer.getOrElse(false)) {
-          // You see review task when:
-          // 1. Project and Challenge enabled (so visible to everyone)
-          // 2. You own the Project
-          // 3. You manage the project (your user group matches groups of project)
-          s"""
+    } else {
+      if (user.settings.isReviewer.getOrElse(false)) {
+        // You see review task when:
+        // 1. Project and Challenge enabled (so visible to everyone)
+        // 2. You own the Project
+        // 3. You manage the project (your user group matches groups of project)
+        s"""
             SELECT ROW_NUMBER() OVER (${orderByClause}) as row_num,
               tasks.${this.retrieveColumnsWithReview} FROM tasks
             ${joinClause}
@@ -327,22 +378,21 @@ class TaskReviewDAL @Inject()(override val db: Database,
             ${orderByClause}
             LIMIT ${sqlLimit(limit)} OFFSET ${offset}
            """
-         }
-         else {
-           return (null, null, null)
-         }
+      }
+      else {
+        return (null, null, null)
+      }
     }
 
-    val countQuery = user.isSuperUser match {
-      case true =>
-        s"""
+    val countQuery = if (user.isSuperUser) {
+      s"""
           SELECT count(*) FROM tasks
           ${joinClause}
           WHERE ${whereClause}
           ${whereBundleClause}
         """
-      case default =>
-        s"""
+    } else {
+      s"""
           SELECT count(*) FROM tasks
           ${joinClause}
           WHERE ((p.enabled AND c.enabled) OR
@@ -355,28 +405,27 @@ class TaskReviewDAL @Inject()(override val db: Database,
         """
     }
 
-    return (countQuery, query, parameters)
+    (countQuery, query, parameters)
   }
 
   /**
     * Gets a list of tasks that have been reviewed (either by this user or requested by this user)
     *
-    * @param user The user executing the request
-    * @param startDate Limit tasks to reviewed after date (YYYY-MM-DD)
-    * @param endDate Limit tasks to reviewed before date (YYYY-MM-DD)
-    * @param asReviewer Whether we should return tasks reviewed by this user or reqested by this user
+    * @param user              The user executing the request
+    * @param startDate         Limit tasks to reviewed after date (YYYY-MM-DD)
+    * @param endDate           Limit tasks to reviewed before date (YYYY-MM-DD)
     * @param allowReviewNeeded Whether we should include review requested tasks as well
-    * @param limit The amount of tasks to be returned
-    * @param offset Offset to start paging
-    * @param sort Column to sort
-    * @param order DESC or ASC
+    * @param limit             The amount of tasks to be returned
+    * @param offset            Offset to start paging
+    * @param sort              Column to sort
+    * @param order             DESC or ASC
     * @return A list of tasks
     */
-  def getReviewedTasks(user:User, searchParameters: SearchParameters,
-                       mappers:Option[List[String]]=None, reviewers:Option[List[String]]=None,
-                       startDate:String, endDate:String, allowReviewNeeded:Boolean=false,
-                       limit:Int = -1, offset:Int=0, sort:String, order:String)
-                       (implicit c:Connection=null) : (Int, List[Task]) = {
+  def getReviewedTasks(user: User, searchParameters: SearchParameters,
+                       mappers: Option[List[String]] = None, reviewers: Option[List[String]] = None,
+                       startDate: String, endDate: String, allowReviewNeeded: Boolean = false,
+                       limit: Int = -1, offset: Int = 0, sort: String, order: String)
+                      (implicit c: Connection = null): (Int, List[Task]) = {
     var orderByClause = ""
     val whereClause = new StringBuilder("(tasks.bundle_id is NULL OR tasks.is_bundle_primary = true) AND ")
     val joinClause = new StringBuilder("INNER JOIN challenges c ON c.id = tasks.parent_id ")
@@ -420,7 +469,7 @@ class TaskReviewDAL @Inject()(override val db: Database,
     val query = user.isSuperUser match {
       case true =>
         s"""
-          SELECT tasks.${this.retrieveColumnsWithReview} FROM tasks
+          SELECT t.${this.retrieveColumnsWithReview} FROM tasks
           ${joinClause}
           WHERE
           ${whereClause}
@@ -447,10 +496,10 @@ class TaskReviewDAL @Inject()(override val db: Database,
             ${orderByClause}
             LIMIT ${sqlLimit(limit)} OFFSET ${offset}
            """
-         }
-         else {
-           return (0, List[Task]())
-         }
+        }
+        else {
+          return (0, List[Task]())
+        }
     }
 
     val countQuery = user.isSuperUser match {
@@ -474,44 +523,110 @@ class TaskReviewDAL @Inject()(override val db: Database,
         """
     }
 
-   var count = 0
-   val tasks = this.taskDAL.cacheManager.withIDListCaching { implicit cachedItems =>
+    var count = 0
+    val tasks = this.manager.task.cacheManager.withIDListCaching { implicit cachedItems =>
       this.withMRTransaction { implicit c =>
         count = sqlWithParameters(countQuery, parameters).as(SqlParser.int("count").single)
         sqlWithParameters(query, parameters).as(this.parser.*)
       }
     }
 
-    return (count, tasks)
+    (count, tasks)
   }
 
+  /**
+    * private setup the search clauses for searching the review tables
+    */
+  private def setupReviewSearchClause(whereClause: StringBuilder, joinClause: StringBuilder,
+                                      searchParameters: SearchParameters,
+                                      startDate: String, endDate: String) {
+    searchParameters.owner match {
+      case Some(o) if o.nonEmpty =>
+        joinClause ++= "INNER JOIN users u ON u.id = tr.review_requested_by "
+        this.appendInWhereClause(whereClause, s"LOWER(u.name) LIKE LOWER('%${o}%')")
+      case _ => // ignore
+    }
+
+    searchParameters.reviewer match {
+      case Some(r) if r.nonEmpty =>
+        joinClause ++= "INNER JOIN users u2 ON u2.id = tr.reviewed_by "
+        this.appendInWhereClause(whereClause, s"LOWER(u2.name) LIKE LOWER('%${r}%')")
+      case _ => // ignore
+    }
+
+    searchParameters.taskStatus match {
+      case Some(statuses) if statuses.nonEmpty =>
+        val statusClause = new StringBuilder(s"(tasks.status IN (${statuses.mkString(",")})")
+        if (statuses.contains(-1)) {
+          statusClause ++= " OR c.status IS NULL"
+        }
+        statusClause ++= ")"
+        this.appendInWhereClause(whereClause, statusClause.toString())
+      case Some(statuses) if statuses.isEmpty => //ignore this scenario
+      case _ =>
+    }
+
+    searchParameters.taskReviewStatus match {
+      case Some(statuses) if statuses.nonEmpty =>
+        val statusClause = new StringBuilder(s"(task_review.review_status IN (${statuses.mkString(",")}))")
+        this.appendInWhereClause(whereClause, statusClause.toString())
+      case Some(statuses) if statuses.isEmpty => //ignore this scenario
+      case _ =>
+    }
+
+    searchParameters.location match {
+      case Some(sl) => this.appendInWhereClause(whereClause, s"tasks.location @ ST_MakeEnvelope (${sl.left}, ${sl.bottom}, ${sl.right}, ${sl.top}, 4326)")
+      case None => // do nothing
+    }
+
+    searchParameters.projectSearch match {
+      case Some(ps) => {
+        val projectName = ps.replace("'", "''")
+        this.appendInWhereClause(whereClause, s"""LOWER(p.display_name) LIKE LOWER('%${projectName}%')""")
+      }
+      case None => // do nothing
+    }
+
+    searchParameters.taskId match {
+      case Some(tid) => this.appendInWhereClause(whereClause, s"CAST(tasks.id AS TEXT) LIKE '${tid}%'")
+      case _ => // do nothing
+    }
+
+    if (startDate != null && startDate.matches("[0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]")) {
+      this.appendInWhereClause(whereClause, "reviewed_at >= '" + startDate + " 00:00:00'")
+    }
+
+    if (endDate != null && endDate.matches("[0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]")) {
+      this.appendInWhereClause(whereClause, "reviewed_at <= '" + endDate + " 23:59:59'")
+    }
+  }
 
   /**
     * Gets a list of tasks that have been reviewed (either by this user or requested by this user)
     *
-    * @param user The user executing the request
+    * @param user      The user executing the request
     * @param reviewTasksType
     * @param searchParameters
     * @param startDate Limit tasks to reviewed after date (YYYY-MM-DD)
-    * @param endDate Limit tasks to reviewed before date (YYYY-MM-DD)
+    * @param endDate   Limit tasks to reviewed before date (YYYY-MM-DD)
     * @return A list of tasks
     */
-  def getReviewMetrics(user:User, reviewTasksType:Int, searchParameters: SearchParameters,
-                       mappers:Option[List[String]]=None, reviewers:Option[List[String]]=None,
-                       priorities:Option[List[Int]]=None, startDate:String, endDate:String,
-                       onlySaved: Boolean=false, excludeOtherReviewers: Boolean=false)
-                       (implicit c:Connection=null) : ReviewMetrics = {
+  def getReviewMetrics(user: User, reviewTasksType: Int, searchParameters: SearchParameters,
+                       mappers: Option[List[String]] = None, reviewers: Option[List[String]] = None,
+                       priorities: Option[List[Int]] = None, startDate: String, endDate: String,
+                       onlySaved: Boolean = false, excludeOtherReviewers: Boolean = false)
+                      (implicit c: Connection = null): ReviewMetrics = {
 
-   // 1: REVIEW_TASKS_TO_BE_REVIEWED = 'tasksToBeReviewed'
-   // 2: MY_REVIEWED_TASKS = 'myReviewedTasks'
-   // 3: REVIEW_TASKS_BY_ME = 'tasksReviewedByMe'
-   // 4: ALL_REVIEWED_TASKS = 'allReviewedTasks'
+    // 1: REVIEW_TASKS_TO_BE_REVIEWED = 'tasksToBeReviewed'
+    // 2: MY_REVIEWED_TASKS = 'myReviewedTasks'
+    // 3: REVIEW_TASKS_BY_ME = 'tasksReviewedByMe'
+    // 4: ALL_REVIEWED_TASKS = 'allReviewedTasks'
 
     val joinClause = new StringBuilder("INNER JOIN challenges c ON c.id = tasks.parent_id ")
     joinClause ++= "LEFT OUTER JOIN task_review ON task_review.task_id = tasks.id "
     joinClause ++= "INNER JOIN projects p ON p.id = c.parent_id "
 
-    var whereClause = new StringBuilder()
+    val whereClause = new StringBuilder()
 
     if (reviewTasksType == 1) {
       this.appendInWhereClause(whereClause,
@@ -547,7 +662,7 @@ class TaskReviewDAL @Inject()(override val db: Database,
                 task_review.reviewed_by = ${user.id})""")
       }
     }
-    
+
     mappers match {
       case Some(m) =>
         if (m.size > 0) {
@@ -587,7 +702,8 @@ class TaskReviewDAL @Inject()(override val db: Database,
     this.appendInWhereClause(whereClause, "(tasks.bundle_id is NULL OR tasks.is_bundle_primary = true)")
     setupReviewSearchClause(whereClause, joinClause, searchParameters, startDate, endDate)
 
-    val query = s"""
+    val query =
+      s"""
      SELECT COUNT(*) AS total,
      COUNT(review_status) FILTER (where review_status = 0) AS requested,
      COUNT(review_status) FILTER (where review_status = 1) AS approved,
@@ -605,7 +721,7 @@ class TaskReviewDAL @Inject()(override val db: Database,
      ${whereClause}
     """
     val reviewMetricsParser: RowParser[ReviewMetrics] = {
-        get[Int]("total") ~
+      get[Int]("total") ~
         get[Int]("requested") ~
         get[Int]("approved") ~
         get[Int]("rejected") ~
@@ -617,9 +733,9 @@ class TaskReviewDAL @Inject()(override val db: Database,
         get[Int]("alreadyFixed") ~
         get[Int]("tooHard") map {
         case total ~ requested ~ approved ~ rejected ~ assisted ~ disputed ~
-             fixed ~ falsePositive ~ skipped ~ alreadyFixed ~ tooHard => {
+          fixed ~ falsePositive ~ skipped ~ alreadyFixed ~ tooHard => {
           new ReviewMetrics(total, requested, approved, rejected, assisted, disputed,
-                            fixed, falsePositive, skipped, alreadyFixed, tooHard)
+            fixed, falsePositive, skipped, alreadyFixed, tooHard)
         }
       }
     }
@@ -641,15 +757,15 @@ class TaskReviewDAL @Inject()(override val db: Database,
     * @param c              an implicit connection
     * @return A list of task clusters
     */
-  def getReviewTaskClusters(user:User, reviewTasksType:Int, params: SearchParameters,
+  def getReviewTaskClusters(user: User, reviewTasksType: Int, params: SearchParameters,
                             numberOfPoints: Int = TaskDAL.DEFAULT_NUMBER_OF_POINTS,
-                            startDate:String, endDate:String, onlySaved: Boolean=false,
-                            excludeOtherReviewers: Boolean=false)
-                     (implicit c: Option[Connection] = None): List[TaskCluster] = {
+                            startDate: String, endDate: String, onlySaved: Boolean = false,
+                            excludeOtherReviewers: Boolean = false)
+                           (implicit c: Option[Connection] = None): List[TaskCluster] = {
     this.withMRConnection { implicit c =>
       val fetchBy = if (reviewTasksType == 2) "task_review.reviewed_by" else "task_review.review_requested_by"
-      val joinClause = new StringBuilder("INNER JOIN challenges c ON c.id = t.parent_id ")
-      joinClause ++= "LEFT OUTER JOIN task_review ON task_review.task_id = t.id "
+      val joinClause = new StringBuilder("INNER JOIN challenges c ON c.id = tasks.parent_id ")
+      joinClause ++= "LEFT OUTER JOIN task_review ON task_review.task_id = tasks.id "
       joinClause ++= "INNER JOIN projects p ON p.id = c.parent_id "
 
       var whereClause = new StringBuilder(s"${fetchBy}=${user.id}")
@@ -675,10 +791,10 @@ class TaskReviewDAL @Inject()(override val db: Database,
                   task_review.review_requested_by != ${user.id} """
         }
       }
-      this.appendInWhereClause(whereClause, "(t.bundle_id is NULL OR t.is_bundle_primary = true)")
+      this.appendInWhereClause(whereClause, "(tasks.bundle_id is NULL OR tasks.is_bundle_primary = true)")
 
       params.taskId match {
-        case Some(tid) => this.appendInWhereClause(whereClause, s"CAST(t.id AS TEXT) LIKE '${tid}%'")
+        case Some(tid) => this.appendInWhereClause(whereClause, s"CAST(tasks.id AS TEXT) LIKE '${tid}%'")
         case _ => // do nothing
       }
 
@@ -689,7 +805,7 @@ class TaskReviewDAL @Inject()(override val db: Database,
         this.appendInWhereClause(whereClause, s"task_review.review_status <> ${Task.REVIEW_STATUS_REQUESTED} ")
       }
 
-      setupReviewSearchClause(whereClause, joinClause, params, startDate, endDate, "t")
+      setupReviewSearchClause(whereClause, joinClause, params, startDate, endDate)
 
       val where = if (whereClause.isEmpty) {
         whereClause.toString
@@ -697,80 +813,136 @@ class TaskReviewDAL @Inject()(override val db: Database,
         "WHERE " + whereClause.toString
       }
 
-      val query = getTaskClusterQuery(joinClause.toString, where, numberOfPoints)
-      var result = sqlWithParameters(query, parameters).as(getTaskClusterParser(params).*)
+      val query = this.manager.taskCluster.getTaskClusterQuery(joinClause.toString, where, numberOfPoints)
+      val result = sqlWithParameters(query, parameters).as(this.manager.taskCluster.getTaskClusterParser(params).*)
       // Filter out invalid clusters.
-      result.filter( _ != None ).asInstanceOf[List[TaskCluster]]
+      result.filter(_ != None).asInstanceOf[List[TaskCluster]]
     }
   }
-
 
   /**
-   * private setup the search clauses for searching the review tables
-   */
-  private def setupReviewSearchClause(whereClause: StringBuilder, joinClause: StringBuilder,
-                                      searchParameters: SearchParameters,
-                                      startDate: String, endDate: String,
-                                      tasksTableName: String = "tasks") {
-    searchParameters.owner match {
-     case Some(o) if o.nonEmpty =>
-       joinClause ++= "INNER JOIN users u ON u.id = task_review.review_requested_by "
-       this.appendInWhereClause(whereClause, s"LOWER(u.name) LIKE LOWER('%${o}%')")
-     case _ => // ignore
+    * Sets the review status for a task. The user cannot set the status of a task unless the object has
+    * been locked by the same user before hand.
+    * Will throw an InvalidException if the task review status cannot be set due to the current review status
+    * Will throw an IllegalAccessException if the user is a not a reviewer, or if the task is locked by
+    * a different user.
+    *
+    * @param task         The task to set the status for
+    * @param reviewStatus The review status to set
+    * @param user         The user setting the status
+    * @return The number of rows updated, should only ever be 1
+    */
+  def setTaskReviewStatus(task: Task, reviewStatus: Int, user: User, actionId: Option[Long], commentContent: String = "")(implicit c: Connection = null): Int = {
+    if (!user.isSuperUser && !user.settings.isReviewer.get && reviewStatus != Task.REVIEW_STATUS_REQUESTED &&
+      reviewStatus != Task.REVIEW_STATUS_DISPUTED) {
+      throw new IllegalAccessException("User must be a reviewer to edit task review status.")
     }
 
-    searchParameters.reviewer match {
-     case Some(r) if r.nonEmpty =>
-       joinClause ++= "INNER JOIN users u2 ON u2.id = task_review.reviewed_by "
-       this.appendInWhereClause(whereClause, s"LOWER(u2.name) LIKE LOWER('%${r}%')")
-     case _ => // ignore
-    }
+    val now = Instant.now()
 
-    searchParameters.taskStatus match {
-      case Some(statuses) if statuses.nonEmpty =>
-        val statusClause = new StringBuilder(s"(${tasksTableName}.status IN (${statuses.mkString(",")})")
-        if (statuses.contains(-1)) {
-          statusClause ++= " OR c.status IS NULL"
-        }
-        statusClause ++= ")"
-        this.appendInWhereClause(whereClause, statusClause.toString())
-      case Some(statuses) if statuses.isEmpty => //ignore this scenario
-      case _ =>
-    }
+    val updatedRows = this.withMRTransaction { implicit c =>
+      var fetchBy = "reviewed_by"
 
-    searchParameters.taskReviewStatus match {
-      case Some(statuses) if statuses.nonEmpty =>
-        val statusClause = new StringBuilder(s"(task_review.review_status IN (${statuses.mkString(",")}))")
-        this.appendInWhereClause(whereClause, statusClause.toString())
-      case Some(statuses) if statuses.isEmpty => //ignore this scenario
-      case _ =>
-    }
+      val isDisputed = task.review.reviewStatus.getOrElse(-1) != Task.REVIEW_STATUS_DISPUTED &&
+        reviewStatus == Task.REVIEW_STATUS_DISPUTED
+      val needsReReview = (task.review.reviewStatus.getOrElse(-1) != Task.REVIEW_STATUS_REQUESTED &&
+        reviewStatus == Task.REVIEW_STATUS_REQUESTED) || isDisputed
 
-    searchParameters.location match {
-      case Some(sl) => this.appendInWhereClause(whereClause, s"${tasksTableName}.location @ ST_MakeEnvelope (${sl.left}, ${sl.bottom}, ${sl.right}, ${sl.top}, 4326)")
-      case None => // do nothing
-    }
+      var reviewedBy = task.review.reviewedBy
+      var reviewRequestedBy = task.review.reviewRequestedBy
 
-    searchParameters.projectSearch match {
-      case Some(ps) => {
-        val projectName = ps.replace("'","''")
-        this.appendInWhereClause(whereClause, s"""LOWER(p.display_name) LIKE LOWER('%${projectName}%')""")
+      // If we are changing the status back to "needsReview" then this task
+      // has been fixed by the mapper and the mapper is requesting review again
+      if (needsReReview) {
+        fetchBy = "review_requested_by"
+        reviewRequestedBy = Some(user.id)
       }
-      case None => // do nothing
+      else {
+        reviewedBy = Some(user.id)
+      }
+      val updatedRows =
+        SQL"""UPDATE task_review SET review_status = $reviewStatus,
+                                 #${fetchBy} = ${user.id},
+                                 reviewed_at = ${now},
+                                 review_started_at = task_review.review_claimed_at,
+                                 review_claimed_at = NULL,
+                                 review_claimed_by = NULL
+                             WHERE task_review.task_id = (
+                                SELECT tasks.id FROM tasks
+                                LEFT JOIN locked l on l.item_id = tasks.id AND l.item_type = ${task.itemType.typeId}
+                                WHERE tasks.id = ${task.id} AND (l.user_id = ${user.id} OR l.user_id IS NULL)
+                              )""".executeUpdate()
+      // if returning 0, then this is because the item is locked by a different user
+      if (updatedRows == 0) {
+        throw new IllegalAccessException(s"Current task [${task.id} is locked by another user, cannot update review status at this time.")
+      }
+
+      // if you set the status successfully on a task you will lose the lock of that task
+      try {
+        this.unlockItem(user, task)
+      } catch {
+        case e: Exception => logger.warn(e.getMessage)
+      }
+
+      webSocketProvider.sendMessage(WebSocketMessages.reviewUpdate(
+        WebSocketMessages.ReviewData(this.getTaskWithReview(task.id))
+      ))
+
+      val comment = commentContent.nonEmpty match {
+        case true => Some(this.manager.comment.add(user, task, commentContent, actionId))
+        case false => None
+      }
+
+      // Don't send a notification for every task in a task bundle, only the
+      // primary task
+      if (task.bundleId.isEmpty || task.isBundlePrimary.getOrElse(false)) {
+        if (!isDisputed) {
+          if (needsReReview) {
+            // Let's note in the task_review_history table that this task needs review again
+            SQL"""INSERT INTO task_review_history
+                              (task_id, requested_by, reviewed_by, review_status, reviewed_at, review_started_at)
+                  VALUES (${task.id}, ${user.id}, ${task.review.reviewedBy},
+                          $reviewStatus, $now, ${task.review.reviewStartedAt})""".executeUpdate()
+            this.manager.notification.createReviewNotification(user, task.review.reviewedBy.getOrElse(-1), reviewStatus, task, comment)
+          }
+          else {
+            // Let's note in the task_review_history table that this task was reviewed
+            SQL"""INSERT INTO task_review_history
+                              (task_id, requested_by, reviewed_by, review_status, reviewed_at, review_started_at)
+                  VALUES (${task.id}, ${task.review.reviewRequestedBy}, ${user.id},
+                          $reviewStatus, $now, ${task.review.reviewStartedAt})""".executeUpdate()
+            this.manager.notification.createReviewNotification(user, task.review.reviewRequestedBy.getOrElse(-1), reviewStatus, task, comment)
+          }
+        }
+        else {
+          // For disputed tasks.
+          SQL"""INSERT INTO task_review_history
+                            (task_id, requested_by, reviewed_by, review_status, reviewed_at, review_started_at)
+                VALUES (${task.id}, ${user.id}, ${task.review.reviewedBy},
+                        $reviewStatus, $now, ${task.review.reviewStartedAt})""".executeUpdate()
+        }
+      }
+
+      this.cacheManager.withOptionCaching { () =>
+        Some(task.copy(review = task.review.copy(
+          reviewStatus = Some(reviewStatus),
+          reviewRequestedBy = reviewRequestedBy,
+          reviewedBy = reviewedBy,
+          reviewedAt = Some(new DateTime()))))
+      }
+
+      if (!needsReReview) {
+        this.manager.user.updateUserScore(None, Option(reviewStatus),
+          task.review.reviewedBy != None, false, task.review.reviewRequestedBy.get)
+      }
+      else if (reviewStatus == Task.REVIEW_STATUS_DISPUTED) {
+        this.manager.user.updateUserScore(None, Option(reviewStatus), true, true, task.review.reviewedBy.get)
+        this.manager.user.updateUserScore(None, Option(reviewStatus), true, false, task.review.reviewRequestedBy.get)
+      }
+
+      updatedRows
     }
 
-    searchParameters.taskId match {
-      case Some(tid) => this.appendInWhereClause(whereClause, s"CAST(${tasksTableName}.id AS TEXT) LIKE '${tid}%'")
-      case _ => // do nothing
-    }
-
-    if (startDate != null && startDate.matches("[0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]")) {
-      this.appendInWhereClause(whereClause, "reviewed_at >= '" + startDate + " 00:00:00'")
-    }
-
-    if (endDate != null && endDate.matches("[0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]")) {
-      this.appendInWhereClause(whereClause, "reviewed_at <= '" + endDate + " 23:59:59'")
-    }
+    updatedRows
   }
-
 }
