+++ app/org/maproulette/models/dal/mixin/Locking.scala
@@ -1,6 +1,4 @@
-// Copyright (C) 2019 MapRoulette contributors (see CONTRIBUTORS.md).
-// Licensed under the Apache License, Version 2.0 (see LICENSE).
-package org.maproulette.models.dal
+package org.maproulette.models.dal.mixin
 
 import java.sql.Connection
 
@@ -8,14 +6,14 @@ import anorm._
 import org.maproulette.data.ItemType
 import org.maproulette.exception.LockedException
 import org.maproulette.models.BaseObject
-import org.maproulette.models.utils.TransactionManager
+import org.maproulette.models.utils.{DALHelper, TransactionManager}
 import org.maproulette.session.User
 
 /**
   * @author mcuthbert
   */
 trait Locking[T <: BaseObject[_]] extends TransactionManager {
-  this: BaseDAL[_, _] =>
+  this: DALHelper =>
 
   /**
     * Unlocks an item in the database
@@ -101,6 +99,33 @@ trait Locking[T <: BaseObject[_]] extends TransactionManager {
   }
 
   /**
+    * Method to lock a single optional item returned in a lambda block. It will first unlock all items
+    * that have been locked by the user
+    *
+    * @param user     The user making the request
+    * @param itemType The type of item that will be locked
+    * @param block    The block of code to execute inbetween unlocking and locking items
+    * @param c        The connection
+    * @return Option object
+    */
+  def withSingleLocking(user: User, itemType: Option[ItemType] = None)(block: () => Option[T])
+                       (implicit c: Option[Connection] = None): Option[T] = {
+    this.withMRTransaction { implicit c =>
+      // if a user is requesting a task, then we can unlock all other tasks for that user, as only a single
+      // task can be locked at a time
+      this.unlockAllItems(user, itemType)
+      val result = block()
+      if (!user.guest) {
+        result match {
+          case Some(r) => lockItem(user, r)
+          case None => // ignore
+        }
+      }
+      result
+    }
+  }
+
+  /**
     * Locks an item in the database.
     *
     * @param user The user requesting the lock
@@ -147,31 +172,4 @@ trait Locking[T <: BaseObject[_]] extends TransactionManager {
           SQL"""DELETE FROM locked WHERE user_id = ${user.id}""".executeUpdate()
       }
     }
-
-  /**
-    * Method to lock a single optional item returned in a lambda block. It will first unlock all items
-    * that have been locked by the user
-    *
-    * @param user     The user making the request
-    * @param itemType The type of item that will be locked
-    * @param block    The block of code to execute inbetween unlocking and locking items
-    * @param c        The connection
-    * @return Option object
-    */
-  def withSingleLocking(user: User, itemType: Option[ItemType] = None)(block: () => Option[T])
-                       (implicit c: Option[Connection] = None): Option[T] = {
-    this.withMRTransaction { implicit c =>
-      // if a user is requesting a task, then we can unlock all other tasks for that user, as only a single
-      // task can be locked at a time
-      this.unlockAllItems(user, itemType)
-      val result = block()
-      if (!user.guest) {
-        result match {
-          case Some(r) => lockItem(user, r)
-          case None => // ignore
-        }
-      }
-      result
-    }
-  }
 }
