+++ app/org/maproulette/models/dal/mixin/TagDALMixin.scala
@@ -1,12 +1,11 @@
-// Copyright (C) 2019 MapRoulette contributors (see CONTRIBUTORS.md).
-// Licensed under the Apache License, Version 2.0 (see LICENSE).
-package org.maproulette.models.dal
+package org.maproulette.models.dal.mixin
 
 import java.sql.Connection
 
 import anorm._
 import org.maproulette.data.{ChallengeType, ItemType, TaskType}
 import org.maproulette.exception.InvalidException
+import org.maproulette.models.dal.{BaseDAL, TagDAL}
 import org.maproulette.models.{BaseObject, Tag}
 import org.maproulette.session.User
