+++ build.sbt
@@ -1,6 +1,6 @@
 name := "MapRouletteAPI"
 
-version := "3.4.0"
+version := "3.7.0"
 
 scalaVersion := "2.12.8"
