+++ conf/evolutions/default/55.sql
@@ -1,6 +1,12 @@
 # --- !Ups
 -- Add task_styles to challenges.
-ALTER TABLE IF EXISTS challenges ADD COLUMN task_styles VARCHAR DEFAULT NULL;;
+SELECT add_drop_column('challenges', 'task_styles', 'VARCHAR DEFAULT NULL');;
+-- add task_id index for status_actions, as it causes deletes to run slower.
+SELECT create_index_if_not_exists('status_actions', 'task_id', '(task_id)');;
+-- Add bundle_id index for when we are updating the bundle id for a task
+SELECT create_index_if_not_exists('tasks', 'bundle_id', '(bundle_id)');;
 
 # --- !Downs
-ALTER TABLE IF EXISTS challenges DROP COLUMN task_styles;;
+SELECT add_drop_column('challenges', 'task_styles', '', false);;
+DROP INDEX IF EXISTS idx_status_actions_task_id;;
+DROP INDEX IF EXISTS idx_tasks_bundle_id;;
